refactor(models): move Counter model from orders.js to counter.js Restructure model files to better organize related code. The Counter model is more appropriately placed in its own dedicated file rather than being in orders.js
